// admin-web/src/types/products/GetAllProductResponse.ts
/**
 * @description Get all product response
 * @author @nhuttan12
 * @version 1.0.0
 * @since 2025-11-12
 */

export interface GetAllProductResponse {
	id: number;
	image: string;
	name: string;
	price: number;
	discount: number;
	color: string;
	rating: number;
	size: string[];
	description: string;
}