// admin-web/src/types/common/SearchField.ts
/**
 * @description Search field props
 * @author @nhuttan12
 * @version 1.0.0
 * @since 2025-11-16
 */

/**
 * @property {string} placeholder - Placeholder of search field
 * @property {boolean} allowClear - Allow clear search field
 * @property {React.CSSProperties} style - Style of search field
 */
export interface SearchField {
	placeholder: string;
	allowClear: boolean;
	style: React.CSSProperties;
}